refactor: restructured contracts into separate modules
Changes:
- Moved SmartContractCounterparty and SmartContractInitiator into separate modules: counterparty_contract.rs and initiator_contract.rs respectively
- Updated imports in blockchain.rs to reflect the new module structure
